+++ src/components/NavBar/Menu.js
@@ -1,0 +1,50 @@
+import Icon from '../Icon/Icon';
+import { COLORS } from '../../utils/constants';
+import {
+  clockIcon,
+  homeIcon,
+  membersIcon,
+  profileIcon,
+} from '../../utils/icons';
+
+export default class Menu {
+  constructor(container) {
+    this.$container = document.querySelector(container);
+    this.menus = [
+      { path: '/', title: '홈', icon: homeIcon },
+      { path: '/members', title: '구성원', icon: membersIcon },
+      { path: '/profile', title: '프로필', icon: profileIcon },
+      { path: '/work-manage', title: '근무/휴가', icon: clockIcon },
+    ];
+    this._active = '/';
+    this.render();
+  }
+
+  set active(activeMenu) {
+    this._active = activeMenu;
+    this.render();
+  }
+
+  render() {
+    this.$container.innerHTML = this.menus
+      .map((menu) => {
+        const isActive = this._active === menu.path;
+        const icon = new Icon({
+          svg: menu.icon,
+          options: {
+            color: isActive ? COLORS.PRIMARY : COLORS.BLACK,
+          },
+        });
+
+        return `
+          <li>
+            <a href="${menu.path}" class="${isActive ? 'active' : ''}">
+              ${icon.html()}
+              <span class="nav-menu">${menu.title}</span>
+            </a>
+          </li>
+        `;
+      })
+      .join('');
+  }
+}
